ai_employee/requirements.txt
# ============================================================
# AI Employee — Python Dependencies
# Gold Tier — Panaversity AI Employee Hackathon 2026
# ============================================================

# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
anthropic>=0.34.0
pydantic>=2.7.0

# Google / Gmail
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# Playwright (WhatsApp)
playwright>=1.44.0

# File watching
watchdog>=4.0.0

# Scheduling
schedule>=1.2.2
apscheduler>=3.10.4

# Social media
tweepy>=4.14.0
requests>=2.32.0
requests-oauthlib>=2.0.0

# Data processing
pandas>=2.2.0

# Utilities
pyyaml>=6.0.1
python-multipart>=0.0.9
websockets>=12.0
httpx>=0.27.0
tenacity>=8.3.0
rich>=13.7.0
typer>=0.12.3
